+++ NEWS.md
@@ -1,3 +1,6 @@
+# Ternary vNEXT
+ - Add function JoinTheDots to plot lines and points simultaneously.
+
 # Ternary v0.1.1
  - Added colour-blind compatible colour palettes `cbPalette8` and `cbPalette15`.
  - Support lab.font in TernaryPlot.